Removing create-react-app from script

+++ install.sh
@@ -8,11 +8,7 @@ ql_no_color='\033[0m'
 
 # Download file function
 downloadfile() {
-    if [[ "$2" == "-f" ]]; then
-        prefix='https://raw.githubusercontent.com/FellowshipOfThePing/js-boilerplate/master/'
-        curl -s -H 'Cache-Control: no-cache' "${prefix}$1" > "./$1"
-        echo -e "${ql_cyan}=> Downloaded $1. ${ql_no_color}";
-    elif [ -f "./$1" ]; then
+    if [ -f "./$1" ]; then
         echo "$1 already exists."
     else 
         prefix='https://raw.githubusercontent.com/FellowshipOfThePing/js-boilerplate/master/'
@@ -26,9 +22,13 @@ promptUser() {
     echo "";
 }
 
-# Create React App
-npx create-react-app .
-promptUser "CRA Succeeded"
+# Create package.json if it doesn't exist
+if [ ! -f ./package.json ]; then
+    npm init -y
+    promptUser "package.json Created"
+else
+    promptUser "package.json already exists"
+fi
 
 # Add template scripts to package.json
 curl -s -H 'Cache-Control: no-cache' https://raw.githubusercontent.com/FellowshipOfThePing/js-boilerplate/master/addScripts.js | node
@@ -56,15 +56,17 @@ downloadfile "CODEOWNERS"
 promptUser "Successfully downloaded files"
 
 # Remove package-lock.json
-rm ./package-lock.json
-promptUser "Removed package-lock.json"
+if [ -f ./package-lock.json ]; then
+    rm ./package-lock.json
+    promptUser "Removed package-lock.json"
+fi
 
 # Add template dependencies to package.json
 yarn add babel-loader patch-package postinstall-postinstall react-router-dom
 yarn add -D @babel/preset-env @babel/preset-react eslint eslint-config-prettier eslint-plugin-import eslint-plugin-jsx-a11y eslint-plugin-prettier eslint-plugin-react prettier
 promptUser "Installed template dependencies"
 
-# Install Yarn Packages
+# Install remaining Yarn packages
 yarn
 
 promptUser "Template download complete!"
